Refactor bot routing into separate controller modules
